[backend] move filter markings logic into markingDefinition utils
